refactor: index file is now exporting
